+++ store/gaskvstore.go
@@ -1,6 +1,8 @@
 package store
 
 import (
+	"io"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -82,7 +84,12 @@ func (gi *gasKVStore) ReverseIterator(start, end []byte) sdk.Iterator {
 
 // Implements KVStore.
 func (gi *gasKVStore) CacheWrap() sdk.CacheWrap {
-	panic("you cannot CacheWrap a GasKVStore")
+	panic("cannot CacheWrap a GasKVStore")
+}
+
+// CacheWrapWithTrace implements the KVStore interface.
+func (gi *gasKVStore) CacheWrapWithTrace(_ io.Writer, _ TraceContext) CacheWrap {
+	panic("cannot CacheWrapWithTrace a GasKVStore")
 }
 
 func (gi *gasKVStore) iterator(start, end []byte, ascending bool) sdk.Iterator {